+++ addon/stores/base-store.js
@@ -10,6 +10,7 @@ import decorateAdapter from './base-store/decorate-adapter';
 import decorateAPICall from './base-store/decorate-api-call';
 import QueryObject from '../query/query-object';
 import isObject from '../utils/is-object';
+import OnlineStore from './online-store';
 
 /**
   Base class for application store.
@@ -175,7 +176,7 @@ export default DS.Store.extend({
 
     // Set online store if it is not specified in application explicitly.
     if (isNone(this.get('onlineStore'))) {
-      let onlineStore = DS.Store.create(owner.ownerInjection());
+      let onlineStore = OnlineStore.create(owner.ownerInjection());
       this.set('onlineStore', onlineStore);
     }
 
@@ -312,6 +313,16 @@ export default DS.Store.extend({
   */
   deleteRecord() {
     return this._callSuperMethod('deleteRecord', 1, arguments);
+  },
+
+  /**
+    Delete all record from the current store.
+    @method deleteRecord
+    @param {DS.Model} record
+    @param {Boolean} [useOnlineStore] Allow to explicitly specify online or offline store using independently of global online status
+  */
+  deleteAllRecords() {
+    return this._callSuperMethod('deleteAllRecords', 2, arguments);
   },
 
   /**